Remove the retry helper cycle
